+++ cmd/vault-plugin-secrets-openldap/main.go
@@ -4,7 +4,7 @@ import (
 	"os"
 
 	hclog "github.com/hashicorp/go-hclog"
-	openldap "github.com/hashicorp/vault-plugin-secrets-openldap"
+	ldap "github.com/hashicorp/vault-plugin-secrets-openldap"
 	"github.com/hashicorp/vault/api"
 	"github.com/hashicorp/vault/sdk/plugin"
 )
@@ -18,7 +18,7 @@ func main() {
 	tlsProviderFunc := api.VaultPluginTLSProvider(tlsConfig)
 
 	err := plugin.Serve(&plugin.ServeOpts{
-		BackendFactoryFunc: openldap.Factory,
+		BackendFactoryFunc: ldap.Factory,
 		TLSProviderFunc:    tlsProviderFunc,
 	})
 	if err != nil {